Extract setting cleaning code to utility class
`Fog.symbolize_credentials` is actually a `Hash.symbolize_keys`
implementation with an exception for Headers.

Whilst it appears to be related to credentials (which is an assortment
of class methods around a Hash not an class) it is only used by
Fog::Service.

This extracts it as a start of a shared utility class whilst various
bits of randomness are pulled to sensible places.