Added feature to overwrite keys on hash of attributes generation
Fog already has support to inform that an attribute might be
initialized with another name. This feature makes possible to map
attributes that aren't ruby friend to friendly ones.

This code makes possible to map the attributes that are ruby friendly
to the methods that the provider need/wants.

Together, the `aliases` and `as` options will map the inialization
of the attributes and the generation of proper attribute names.