Fix alternative require ordering
Earlier providers tended to have modules named `Fog::Service::Provider`
but defined in a file `fog/provider/service`

Whilst the `Fog::Compute` module dealt with the alternative namespace
(`Fog::Provider::Service`) it did not match with the clearer file naming

This rescues the `LoadError` and attempts the alternative form.

This code is all due for removal real soon now.