[GH105] Start to clean up Fog::Service code
This adds a `ServiceAbstraction` module that can be used to extend a
`Fog::Service` module such as Compute or Storage.

Also included is a testing spec module for use in `fog` itself to test
provider specific cases. These cases will be eliminate in core but will
still require compatibility testing initially.