[GH105] Use generic code for service abstraction
Clearing out the copypasta, and handling the two require and two class
name combinations allows us to eliminate most of the provider specific
rubbish.

Version abstraction options have also been disabled.

The remaining errors appear to be related to providers who changed the
form of their names. To compensate there is a little look up table to
correct this for the moment.