Simpler check for Ruby version and skip
This removes the need for Rubygems just to parse the `RUBY_VERSION`
string to ignore 1.8.7

String comparison gives us a simpler way to check this and I started
seeing errors about frozen Strings being modified related to this.

This now "skips" on any ruby < 1.9. string to ignore 1.8.7

String comparison gives us a simpler way to check this and I started
seeing errors about frozen Strings being modified related to this.

This now "skips" on any ruby < 1.9.3