Only rescue errors related to loading and finding components
Allow connection errors to bubble up to the caller.  Previous changes were
rescuing connection errors and re-raising them as ArgumentErrors leading
people to believe that "openstack has no network service" for example.
However, this was not the case, openstack does have a network service
registered in fog, the proper library was loaded, but there was an issue
connecting to the server.

See changes in:
https://github.com/fog/fog-core/commit/36e2824c8467402697edc188cc8779ced659e31c