Add exponential increase on timeout in #sshable? for services that do not respond in 8s
Issue fog/fog-aws#372

For AWS Spot Request, instances never complete the `#setup` process
eventually timing out through the `#wait_for` block because the
default 8s is apparently not long enough to get the ssh connection.
Through testing 11s seemed to work, but this is likely highly variable
for regions, instance types, images, and providers.

This change implements a 1.5 increase in timeout each time #sshable?
is called, capped at 60s. If a successful connection is made, then
the timeout is reset to the initial value.