Log issues when loading providers
I was hitting issue Fog::Service::NotFound: ovirt has no compute service
when trying to use fog-ovirt and there was no indication of what could be
wrong. Eventually, I've figured out, that there was issue with
dependencies inside fog-ovirt, that we were suppressing.

This patch adds logging around the requiring the provider so that this
issues are easier to debug later.